+++ src/mine/mine_manager.cpp
@@ -333,7 +333,7 @@ void MineManager::generateStats(int64_t age_in_days) {
         ProgramUtil::removeOps(program, Operation::Type::NOP);
 
         // update stats
-        stats->updateProgramStats(s.id, program, submitter);
+        stats->updateProgramStats(s.id, program, submitter, has_formula);
         num_processed++;
       } catch (const std::exception &exc) {
         Log::get().error(